flash_writer_tool: wait before writing FLASHWRITER
RZ/T2H and RZ/N2H platforms require a special header to be written
before the FLASHWRITER.

Give SCI a chance to handle the special header and to put itself into
program loading mode by waiting for 1 second.

The maximum delay for the FLASHWRITER file to start being written is 10
seconds, as described in the BSP docs.